+++ EasyStock.Domain/Services/CalculadoraResumoVenda.cs
@@ -7,20 +7,20 @@ namespace EasyStock.Domain.Services
 {
     public sealed class CalculadoraResumoVenda
     {
-        public record Resumo(decimal ValorTotal, int QuantidadeTotal);
+        public record Resumo(decimal ValorTotal, decimal QuantidadeTotal);
 
         // Recebe itens com quantidade e preco unitario e calcula um resumo simples
         public Resumo Calcular(IEnumerable<(Quantidade quantidade, Dinheiro precoUnitario)> itens)
         {
             if (itens == null) throw new ArgumentNullException(nameof(itens));
-            long somaQuantidade = 0;
+            decimal somaQuantidade = 0m;
             decimal somaValor = 0m;
             foreach (var (quantidade, precoUnitario) in itens)
             {
                 somaQuantidade += quantidade.Value;
                 somaValor += precoUnitario.Valor * quantidade.Value;
             }
-            return new Resumo(somaValor, (int)somaQuantidade);
+            return new Resumo(somaValor, somaQuantidade);
         }
     }
 }
